Functional PCA code
This code was implemented by all team members (data wrangling). However, Eden and I were working on FPCA. She added the base code for this technique and I include other parts to answer the most recent question from our last meeting with the client about which waveform influence the behavior of the second component displayed by Sensor A - CalWindow. So, the personal improvements on the code are the following:
1. Make the FPCA into a function to facilitate and make less verbose the code.
2. Add traditional PCA to compare the plots of the principal components versus those return by FPCA.
3. Return the scores to identify the TestID of the waveforms that accounts of the greater variability as required by the client.
4. Filter and identify the additional features of the TestID identified in 3.